switch tests to refactored code
