--- tests/unit/pipeline/ingest/test_ingest_refactor.py
@@ -1,111 +1,0 @@
-# tests for optimised ingest
-# this is termporary and will go away once optised version is merged
-
-import unittest, unittest.mock
-from unittest.mock import patch
-
-import signal
-import subprocess
-#import pickle
-import context
-import ingest_refactor as ingest
-
-
-test_alert = {
-    'DiaObject': { 'diaObjectId': 1998903343203749723, },
-    'DiaSource': { 'diaSourceId': 181071530527032103, 'midPointTai': 57095.171263959775 },
-    'DiaSourceList': [ { 'diaSourceId': 176546782480695886, 'midPointTai': 57070.34313563427 }, { 'diaSourceId': 176891668354564641, 'midPointTai': 57072.3425344742 } ],
-    'ForcedSourceOnDiaObjectList': [ {}, {} ],
-    'cutoutDifference': b'foo',
-    'cutoutTemplate': b'bar',
-}
-
-class IngestOptTest(unittest.TestCase):
-
-    # check that the sigterm handler sets sigterm raised correctly
-    def test_sigterm_handler(self):
-        ingester = ingest.Ingester('', '', '', 1)
-        self.assertFalse(ingester.sigterm_raised)
-        subprocess.run(['pkill', '-f', 'python3 test_ingest_refactor.py'])
-        self.assertTrue(ingester.sigterm_raised)
-
-    # test using the image store 
-    def test_store_images(self):
-        mock_image_store = unittest.mock.MagicMock()
-        diaSourceId = test_alert['DiaSource']['diaSourceId']
-        diaObjectId = test_alert['DiaObject']['diaObjectId']
-        imjd = int(test_alert['DiaSource']['midPointTai'])
-        imageStore = ingest.ImageStore(image_store=mock_image_store)
-        result = imageStore.store_images(test_alert, diaSourceId, imjd, diaObjectId)
-        # check we called putCutoutAsync twice
-        self.assertEqual(mock_image_store.putCutoutAsync.call_count, 2)
-
-    @patch('ingest_refactor.ImageStore')
-    @patch('ingest_refactor.Cluster')
-    @patch('ingest_refactor.Producer')
-    @patch('ingest_refactor.Consumer')
-    def test_setup(self,
-            mock_consumer,
-            mock_producer,
-            mock_cluster,
-            mock_image_store):
-        ingester = ingest.Ingester('', '', '', 1)
-        ingester.setup()
-        self.assertEqual(mock_image_store.call_count, 1)     
-        self.assertEqual(mock_cluster.call_count, 1)     
-        self.assertEqual(mock_cluster.return_value.connect.call_count, 1)     
-        self.assertEqual(mock_cluster.return_value.connect.return_value.set_keyspace.call_count, 1)     
-        self.assertEqual(mock_producer.call_count, 1)     
-        self.assertEqual(mock_consumer.call_count, 1)     
-        self.assertEqual(mock_consumer.return_value.subscribe.call_count, 1)     
-
-    @patch('ingest_refactor.executeLoadAsync')
-    def test_insert_cassandra(self, mock_executeLoadAsync):
-        alert = {
-            'diaObject':                 test_alert['DiaObject'],
-            'diaSourcesList':            test_alert['DiaSourceList'],
-            'forcedSourceOnDiaObjectsList':      test_alert['ForcedSourceOnDiaObjectList'],
-        }
-        ingester = ingest.Ingester('', '', '', 1, cassandra_session=True)
-        ingester._insert_cassandra(alert)
-        # executeLoad should get called three times, once for diaObject and once for each list
-        self.assertEqual(mock_executeLoadAsync.call_count, 3)     
-
-    @patch('ingest_refactor.Ingester._insert_cassandra_multi')
-    def test_handle_alert(self, mock_insert_cassandra_multi):
-        mock_insert_cassandra_multi.return_value = [ unittest.mock.MagicMock() ]
-        mock_image_store = unittest.mock.MagicMock() 
-        mock_producer = unittest.mock.MagicMock()
-        ingester = ingest.Ingester('', '', '', 1, image_store=mock_image_store, producer=mock_producer)
-        result = ingester._handle_alert(test_alert)
-        # check the return values
-        self.assertEqual(result, (2, 2))
-        # store_images should get called once
-        mock_image_store.store_images.assert_called_once()
-        # insert_cassandra_multi should get called once
-        mock_insert_cassandra_multi.assert_called_once()
-        # producer.produce should get called once
-        mock_producer.produce.assert_called_once()
-    
-    def test_end_batch(self):
-        mock_log = unittest.mock.MagicMock()
-        mock_consumer = unittest.mock.MagicMock()
-        mock_producer = unittest.mock.MagicMock()
-        mock_ms = unittest.mock.MagicMock()
-        ingester = ingest.Ingester('', '', '', 1, log=mock_log, producer=mock_producer, consumer=mock_consumer)
-        ingester._end_batch(mock_ms, 1, 2, 2)
-        # log message should get sent
-        mock_log.info.assert_called_once()
-        # producer should get flushed
-        mock_producer.flush.assert_called_once()
-        # consumer offsets should get committed
-        mock_consumer.commit.assert_called_once()
-        # status page should get updated
-        mock_ms.add.assert_called_once()
-
-
-if __name__ == '__main__':
-    import xmlrunner 
-    runner = xmlrunner.XMLTestRunner(output='test-reports')
-    unittest.main(testRunner=runner)
-    unittest.main()
